Abstract notification building to base class
