+++ notifications/update_favorites.py
@@ -6,6 +6,8 @@ from notifications.base_notification import BaseNotification
 
 class UpdateFavoritesNotification(BaseNotification):
 
+    _supported_clients = [ClientType.OS_ANDROID]  # We don't want to send these updates to webhooks
+
     def __init__(self, user_id, sending_device_key):
         self.user_id = user_id
         self.sending_device_key = sending_device_key
